Support for case-sensitive file systems
Added searching for PERSONALITIES.json with a case-insensitive regex against all files in a mod folder and using the first match instead
when the expected "PERSONALITIES.json" could not be found.

For example because it was named "personalities.json" which makes a difference on case-sensitive file systems (e.g. on SteamOS/Linux)